test: Added end-to-end tests for encryption using IV
